Fix MacOS compiler warnings, some of which are real
git-svn-id: svn+ssh://svn.code.sf.net/p/x3270/x3270/trunk@1969 12e4f418-770b-42d6-82dd-bd6eefd7344d

+++ x3270/x3270/save.c
@@ -374,7 +374,7 @@ save_toggles(void)
     int ix;
 
     for (i = 0; i < N_TOGGLES; i++) {
-	if (toggle_names[i].index < 0 || !toggle_changed(i)) {
+	if ((int)toggle_names[i].index < 0 || !toggle_changed(i)) {
 	    continue;
 	}
 
@@ -653,7 +653,7 @@ save_options(char *n)
 
     /* Save most of the toggles. */
     for (i = 0; i < N_TOGGLES; i++) {
-	if (toggle_names[i].index < 0 || !toggle_changed(i)) {
+	if ((int)toggle_names[i].index < 0 || !toggle_changed(i)) {
 	    continue;
 	}
 	if (i == TRACING || i == SCREEN_TRACE) {
